==> crater_env/notebooks/02_find_new_playlists.ipynb <==
from mb_functions import *

import json

SPOTIPY_CLIENT_ID= 'a5d9928d7d874cebafc8a8a632277854'
SPOTIPY_CLIENT_SECRET= 'a3341b047daa4007a5bc3ad2fc9413c7'
SPOTIPY_REDIRECT_URI = 'http://127.0.0.1:9090'
SCOPE = 'user-top-read user-library-read playlist-modify-public'

auth_manager=SpotifyOAuth(client_id=SPOTIPY_CLIENT_ID, client_secret=SPOTIPY_CLIENT_SECRET, redirect_uri=SPOTIPY_REDIRECT_URI, scope=SCOPE)

token = auth_manager.get_access_token()['access_token']

playlist_names = pd.read_csv('assets/playlists_for_search.csv')

remove = ['dad','playlist','new playlist 6','november 2015','hi','jan 2016','october 2014',
          'fall 2016','summa','songs i like',"spring '16","other",'september 2017','lol',
          'my favs','summer 2016','winter 2016','all','new new','august 2015','fall 2014',
          'main',"old school mix",'new']
          
add = ['main character','sit and think','gloomy','warm hug','underground',
       'Simulation','happy but also sad','hopeless','rebellious','fall in love',
       'rainy day','toykyo city pop','pop punk','hiphop minimialism','small artists',
       'feel sexy','songs that make me cry','smoke break','going for a walk','demon time',
       'skating','skateboard','dreamy','levatating','coming of age','villian',"don't give a fuck"]

playlist_names = playlist_names[~playlist_names.names.isin(remove)]

removed_list = list(playlist_names.names)
removed_list.extend(add)

playlist_names = pd.DataFrame(removed_list, columns = ['names'])

## Open AI API Call to Filter non thematic playlists (Not implemented yet)

# import openai

# def filter_playlists_with_openai(playlists_dict, search_input,api_key):
#     """
#     Filters playlists by asking OpenAI to determine thematic consistency among playlist names,
#     using the specified OpenAI Chat API call format.
    
#     Args:
#         playlists_dict (dict): A dictionary of playlist IDs and names.
#         api_key (str): Your OpenAI API key.

#     Returns:
#         dict: A filtered dictionary of playlists that are thematically consistent.
#     """
#     # Initialize the OpenAI client with the provided API key
#     client = openai.OpenAI(api_key=api_key)
    
#     try:
        
#         thread = client.beta.threads.create()
        
#         message = client.beta.threads.messages.create(
#             thread_id=thread.id,
#             role="user",
#             content=f"{playlists_dict}" + f"central term: {search_input}"
#         )
        
#         run = client.beta.threads.runs.create(
#             thread_id=thread.id,
#             assistant_id= "asst_7IK9iOrv80KHehDWRvjq1uju" 
#         )
#         messages = client.beta.threads.messages.list(thread_id=thread.id)
#         print(messages.data[0].content[0].text)
#         output = eval(messages.data[0].content[0].text.value)
        
#         return output
#     except Exception as e:
#         print(f"Failed to query OpenAI: {e}")
#         return {}

# # client = openai.OpenAI()
# api_key = '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'
# client = openai.OpenAI(api_key=api_key)

# thread = client.beta.threads.create()

# message = client.beta.threads.messages.create(
#     thread_id=thread.id,
#     role="user",
#     content=f"{playlist_names}" + f"central term: {search_input}"
# )

# run = client.beta.threads.runs.create(
#     thread_id=thread.id,
#     assistant_id= "asst_7IK9iOrv80KHehDWRvjq1uju" 
# )

import requests
import pandas as pd

# Assuming filter_playlists_with_openai is defined as previously discussed, including the import of requests

def search_playlists(search_input, access_token, api_key):
    headers = {
        'Authorization': f'Bearer {access_token}'
    }
    
    search_url = f'https://api.spotify.com/v1/search?q={search_input}&type=playlist&limit=15'
    response = requests.get(search_url, headers=headers)
    print(response)
    playlists = response.json()['playlists']['items']
    
    if not playlists:
        print("No playlists found for the given search input.")
        return pd.DataFrame()  # Return an empty data frame if no playlists found
    
    playlists_dict = {playlist['id']: playlist['name'] for playlist in playlists}
    # for name in playlists_dict.values():
    
    # # Filter playlists using OpenAI
    # print(playlists_dict)
    # filtered_playlists_dict = filter_playlists_with_openai(playlists_dict, search_input, api_key)
    # print(filtered_playlists_dict)
    # print("\nPlaylists after filtering:")
    # for name in filtered_playlists_dict.values():
    #     print(name)
    
    data = []
    playlist_number = 1
    for playlist_id, playlist_name in playlists_dict.items():
        playlist_url = f'https://api.spotify.com/v1/playlists/{playlist_id}/tracks'
        tracks_response = requests.get(playlist_url, headers=headers)
        tracks_data = tracks_response.json().get('items')
        
        for track in tracks_data:
            if track and track.get('track'):
                track_id = track['track']['id']
                track_name = track['track']['name']
                data.append({
                    'search_id': search_input,
                    'playlist_number': playlist_number,
                    'playlist_id': playlist_id,
                    'playlist_name': playlist_name,
                    'track_id': track_id,
                    'track_name': track_name
                })
        playlist_number += 1  # Increment for each playlist
    
    return pd.DataFrame(data)



import time

# all_data = pd.DataFrame()

all_data

# Assuming `playlist_names` is a DataFrame with a column 'names' that contains playlist names.
# Also assuming `access_token` has been obtained as previously described.

count = 1
for row in list(playlist_names.names[220:]):
    print(count)
    search_input = row
    df_temp = search_playlists(search_input, token,api_key = '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')
    all_data = pd.concat([all_data, df_temp], ignore_index=True)
    time.sleep(1)
    count +=1


song_count_per_search = all_data.groupby(['search_id', 'track_id']).size().reset_index(name='song_count')


all_data_new = all_data.merge(song_count_per_search, on=['search_id', 'track_id'], how='left')


all_data_filtered = all_data_new[all_data_new.song_count >= 2]

all_data_filtered_no_dupe = all_data_filtered.drop_duplicates(subset = ['search_id','track_id'])

all_data_filtered_no_dupe.to_csv('assets/tracks_vibes.csv', index = False)
